[ref #71] member登録時に他メンバーの承認が要るようにする
participantモデルの認証周りのメソッドをまとめた
